server/requirements.txt
Django==1.5
Pillow==2.0.0
boto==2.8.0
distribute==0.6.31
dj-database-url==0.2.1
django-tastypie==0.9.14
gunicorn==0.17.2
mimeparse==0.1.3
psycopg2==2.4.6
python-dateutil==2.1
raven==3.3.2
six==1.3.0
wsgiref==0.1.2
